% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.twoway.R
\name{anova.twoway}
\alias{anova.twoway}
\title{ANOVA summary for a two-way table, including Tukey Additivity Test}
\usage{
\method{anova}{twoway}(object, ...)
}
\arguments{
\item{object}{a \code{class("twoway")} object}

\item{...}{other arguments passed down, but not used here}
}
\description{
Test for a 1-df interaction in two-way ANOVA table by the Tukey test.
}
\details{
At present, this function simply gives the results of the ANOVAs for the additive model, the model including the 1 df
         term for non-additivity, and an \code{anova()} comparison of the two.
         The analysis is based on row and column means.
}
\examples{
data(sentRT)
sent.2way <- twoway(sentRT)
anova(sent.2way)

data(EastCoast)
EC.2way <- twoway(EastCoast)
anova(EC.2way)

data(hstart)
hstart.2way <- twoway(hstart)
anova(hstart.2way)

data(Arizona)
AR.2way <- twoway(Arizona)
anova(AR.2way)
}
\author{
Michael Friendly
}
