% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{drugs}
\alias{drugs}
\title{Scores for 5 subjects after being given each of 4 drugs}
\description{
The original source is Winer (1971), p. 268.  This was used as an example in Friendly (1991).
}
\examples{
data(drugs)
tw <- twoway(drugs) |>
  print()

plot(tw)
}
\references{
Friendly, M. (1991). \emph{SAS System for Statistical Graphics} Cary, NC: SAS Institute, Output 7.28
}
\keyword{data}
