% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hstart}
\alias{hstart}
\title{Number of U.S. housing starts by month for the years 1965 -- 1973}
\format{
a 9 x 12 matrix, where the entries are the number of housing starts, in thousands
}
\description{
Number of U.S. housing starts by month for the years 1965 -- 1973
}
\examples{
hstart.2way <- twoway(hstart, method="mean")
anova(hstart.2way)
plot(hstart.2way)
}
\references{
Becker, Chambers & Wilks (1988), \emph{The New S Language}, Brooks Cole.
            Friendly, M. (1991). \emph{SAS System for Statistical Graphics} Cary, NC: SAS Institute, p.380
}
\keyword{data}
