% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medianfit.R
\name{meanfit}
\alias{meanfit}
\title{Fit a two-way table using row and column means}
\usage{
meanfit(x, ...)
}
\arguments{
\item{x}{a numeric matrix}

\item{...}{other arguments passed down}
}
\value{
An object of class \code{c("twoway")} with the following named components:
\describe{
 \item{overall}{the fitted constant term.}
 \item{row}{the fitted row effects.}
 \item{col}{the fitted column effects.}
 \item{residuals}{the residuals.}
 \item{name}{the name of the dataset.}
}
}
\description{
Fit a two-way table using row and column means
}
