% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medianfit.R
\name{medianfit}
\alias{medianfit}
\title{Fit a two-way table using median polish}
\usage{
medianfit(x, trace.iter = FALSE, ...)
}
\arguments{
\item{x}{a numeric matrix}

\item{trace.iter}{whether to give verbose output of iteration history in median polish.}

\item{...}{other arguments passed down}
}
\value{
An object of class \code{c("twoway", "medpolish")} with the following named components:
\describe{
 \item{overall}{the fitted constant term.}
 \item{row}{the fitted row effects.}
 \item{col}{the fitted column effects.}
 \item{residuals}{the residuals.}
 \item{name}{the name of the dataset.}
}
}
\description{
Fit a two-way table using median polish
}
