% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Rubber}
\alias{Rubber}
\title{Compressibility of Rubber}
\format{
a 4 x 6 matrix, where the cell values are the specific volume (in cubic centimeters per gram)
        of peroxide-cured rubber. The row and column variables are:
 \itemize{
     \item{Temperature}, in degrees Celsius
     \item{Pressure}, in kg / cm^2 above atmospheric pressure.
 }
}
\source{
Wood, L. A. & Martin, G. M. (1964). "Compressibility of natural rubber at pressures below
     500kg/cm^2",  Journal of Research of the National Standards Bureau--A. Physics & Chemistry,
     **68A**, 259--268.
}
\usage{
Rubber
}
\description{
The specific volume of natural rubber was measured at four values of temperature and six values of pressure.
Is there any evidence that volume is not an additive relation with temperature and pressure?
}
\examples{
data(Rubber)
# scale the response to avoid small decimals
rub <- 10000*Rubber
rubfit <- twoway(rub, "median")
plot(rubfit)
}
\references{
Emerson, J. D. & Wong, G. Y. (1985). "Resistant Nonadditve Fits for Two-Way Tables".
     In Hoaglin, D. C., Mosteller, F., & Tukey, J. W. (Eds.). Exploring data tables, trends and shapes. John Wiley Sons.
     Ch. 3, Table 3.1.
}
\keyword{data}
