% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sentRT}
\alias{sentRT}
\title{Reaction times for T/F judgments}
\description{
A demonstration 3 x 3 two-way table composed of reaction times for
three subjects making T/F judgments on three types of sentences
}
\examples{
data(sentRT)
twoway(sentRT)
}
\references{
Friendly, M. (1991). \emph{SAS System for Statistical Graphics} Cary, NC: SAS Institute, Table 7.2
}
\keyword{data}
