% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{taskRT}
\alias{taskRT}
\title{Data on reaction times for various tasks and topics}
\description{
A demonstration 3 x 4 two-way table composed of reaction times for
tasks varying in difficulty, with content on different topics.
}
\keyword{data}
