% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{VermontPop}
\alias{VermontPop}
\title{Vermont country populations from the US Census, 1900-1990}
\format{
An object of class \code{data.frame} with 14 rows and 10 columns.
}
\usage{
VermontPop
}
\description{
Vermont country populations from the US Census, 1900-1990
}
\examples{
options(digits=4)
VP <- twoway(VermontPop,
             method="median",
             responseName = "log Population")
VP
plot(VP)
}
\references{
Morgenthaler, Stephan, and John W. Tukey. “Multipolishing and Two-Way Plots.” Metrika 53.3 (2001): 245–267.
}
\keyword{datasets}
