% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.twoway.R
\name{anova.twoway}
\alias{anova.twoway}
\title{ANOVA summary for a two-way table, including Tukey Additivity Test}
\usage{
\method{anova}{twoway}(x, ...)
}
\arguments{
\item{x}{a \code{class("twoway")} object}

\item{...}{other arguments passed down}
}
\description{
Test for a 1-df interaction in two-way ANOVA table by the Tukey test.
}
