% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Arizona}
\alias{Arizona}
\title{Mean monthly temperatures in Arizona}
\description{
Mean monthly temperatures in Arizona
}
\examples{
data(Arizona)
twoway(Arizona)
plot(twoway(Arizona))
}
\references{
Tukey, J. W. (1977). \emph{Exploratory Data Analysis}, Reading MA: Addison-Wesley. Exhibit 1 of chapter 10, p. 333
}
\keyword{data}
