% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Arizona}
\alias{Arizona}
\title{Mean monthly temperatures in Arizona}
\format{
a matrix of 7 rows (Month) and 3 columns (City) where the value is mean
      monthly temperature in degrees F. The matrix has a \code{responseName} attribute, \code{"Temperature"}
}
\description{
This is the data set used by Tukey (1977) for the initial examples of twoway tables
}
\examples{
data(Arizona)
(AR.2way <-twoway(Arizona, method="median"))

## plot(AR.2way)
}
\references{
Tukey, J. W. (1977). \emph{Exploratory Data Analysis}, Reading MA: Addison-Wesley. Exhibit 1 of chapter 10, p. 333
}
\keyword{data}
