% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medianfit.R
\name{meanfit}
\alias{meanfit}
\title{Fit a two-way table using row and column means}
\usage{
meanfit(x, ..., na.rm = FALSE)
}
\arguments{
\item{x}{a numeric matrix or data frame}

\item{...}{other arguments passed down}

\item{na.rm}{logical. Should missing values be removed?}
}
\value{
An object of class \code{c("twoway")} with the following named components:
\describe{
 \item{overall}{the fitted constant term.}
 \item{roweff}{the fitted row effects.}
 \item{coleff}{the fitted column effects.}
 \item{residuals}{the residuals.}
 \item{name}{the name of the dataset.}
 \item{rownames}{the names for the rows}
 \item{colnames}{the names for the columns}
 \item{method}{"median"}
}
}
\description{
Fit a two-way table using row and column means
}
