% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoway.R
\name{plot.twoway}
\alias{plot.twoway}
\title{Plot method for two-way tables}
\usage{
\method{plot}{twoway}(x, which = c("fit", "diagnose"), main, ylab,
  annotate = TRUE, rfactor = 1, rcolor = c("blue", "red"), ...)
}
\arguments{
\item{x}{a \code{class("twoway")} object}

\item{which}{one of \code{"fit"} or \code{"diagnose"}}

\item{main}{plot title}

\item{ylab}{Y axis label for \code{"fit"} plot}

\item{annotate}{A logical value; if \code{TRUE}, the slope and power are displayed in the diagnostic plot}

\item{rfactor}{for the \code{"fit"} method, draw lines for \code{abs(residuals) > rfactor*sqrt(MSPE)}}

\item{rcolor}{for the \code{"fit"} method, a vector of length 2 giving the color of lines for positive and negative residuals}

\item{...}{other arguments passed down}
}
\description{
Plots either the fitted values and residuals under additivity or
a diagnostic plot for removable non-additivity by a power transformation
}
\details{
For the \code{which="fit"} plot, the basic result comes from a plot of the row effects against the column fitted
    values, which appears as a rectangular grid in these coordinates.  Rotating this 45 degrees counterclockwise give a plot
    in which the vertical coordinate is the fitted value for the two-way table, and the horizontal coordinate is the column fit
    minus the row effect.  The spacing of the grid lines for the rows and columns of the table show the relative magnitudes of the
    row/column means or medians.

    For the \code{which="diagnose"} plot, the interaction residuals from an additive model, \eqn{y_{ij} = \mu + \alpha_i + \beta_j},
    are plotted against the estimated components \eqn{\alpha_i \beta_j / \mu}. If this plot shows a substantially non-zero
    slope, \eqn{b}, this analysis suggests that a power transformation, \eqn{ y \rightarrow y^(1-b)} might reduce the
    apparent interaction effects.
}
\examples{
data(sentRT)
twoway(sentRT)
plot(twoway(sentRT), ylab="Reaction Time")
plot(twoway(sentRT), which="diagnose")
}
\author{
Michael Friendly
}
