% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoway.R
\name{print.twoway}
\alias{print.twoway}
\title{Print method for two-way tables}
\usage{
\method{print}{twoway}(x, digits = getOption("digits"), ...)
}
\arguments{
\item{x}{a numeric matrix}

\item{digits}{number of digits to print}

\item{...}{other arguments passed down}
}
\description{
Print method for two-way tables
}
