% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoway.formula.R
\name{twoway.formula}
\alias{twoway.formula}
\title{Initial sketch for a twoway formula method}
\usage{
\method{twoway}{formula}(formula, data, subset, na.action, ...)
}
\arguments{
\item{formula}{A formula of the form \code{response ~ rowvar + colVAR}}

\item{data}{The name of the data set}

\item{subset}{An expression to subset the data}

\item{na.action}{What to do with NAs?}

\item{...}{other arguments, passed down}
}
\description{
Doesn't do anything useful yet, but the idea is to be able to use a
formula for a twoway table in long form, e.g.,
twoway(response ~ row + col, data=mydata)
}
