% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoway.R
\name{twoway}
\alias{twoway}
\alias{twoway.default}
\title{Analysis of a two-way table}
\usage{
twoway(x, ...)

\method{twoway}{default}(x, method = c("mean", "median"), ...)
}
\arguments{
\item{x}{a numeric matrix or data frame}

\item{...}{other arguments passed down}

\item{method}{one of \code{"mean"} or \code{"median"}}
}
\value{
An object of class \code{"twoway"}
}
\description{
Fits an additive model using either row and column means or Tukey's median polish procedure

Default method for two-way tables
}
\details{
The \code{rownames(x)} are used as the levels of the row factor and the \code{colnames(x}) are
         the levels of the column factor.
}
\examples{
data(taskRT)
twoway(taskRT)

}
\references{
Tukey, J. W. (1977). \emph{Exploratory Data Analysis}, Reading MA: Addison-Wesley.
            Friendly, M. (1991). \emph{SAS System for Statistical Graphics} Cary, NC: SAS Institute
}
\seealso{
code{\link[stats]{medpolish}}

\code{\link{medianfit}}, \code{\link{meanfit}}
}
\author{
Michael Friendly
}
